// frontend/src/types/api.ts
export interface components {
    schemas: {
// ...
        InventoryEntryResult: {
            /** Format: uuid */
            id?: string;
            item?: components["schemas"]["ItemResult"];
            /** Format: int32 */
            quantity?: number;
            /** Format: date-time */
            addedAt?: string;
        };
        InventoryResult: {
            /** Format: uuid */
            adventureId?: string;
            entries?: components["schemas"]["InventoryEntryResult"][] | null;
            /** Format: int32 */
            totalEntries?: number;
        };
        ItemListResult: {
            items?: components["schemas"]["ItemResult"][] | null;
            /** Format: int32 */
            totalCount?: number;
        };
        ItemResult: {
            /** Format: uuid */
            id?: string;
            name?: string | null;
            description?: string | null;
            rarity?: string | null;
            itemType?: string | null;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}